requirements.txt
google-api-python-client==2.90.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.3
oauthlib==3.2.2
httplib2==0.21.0
requests==2.31.0
urllib3==2.0.7
